fix: fazendo a conversão correta de DTO no service de buscarPorInicioDoNome

+++ ms-nutria-mongodb/src/main/java/org/example/msnutriamongodb/repository/IngredienteRepository.java
@@ -7,10 +7,9 @@ import org.springframework.data.mongodb.repository.Aggregation;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
-import java.util.regex.Pattern;
 
 @Repository
-public interface IngredienteRepository extends MongoRepository<Ingrediente, Long> {
+public interface IngredienteRepository extends MongoRepository<Ingrediente, Integer> {
   Boolean existsByNomeIngrediente(String nomeIngrediente);
 
   Optional<Ingrediente> findById(Integer id);
